requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.11.1
bleach==5.0.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.0
charset-normalizer==3.1.0
chromadb==0.3.21
click==8.1.3
clickhouse-connect==0.5.23
contourpy==1.0.7
cycler==0.11.0
dataclasses-json==0.5.7
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
duckdb==0.7.1
entrypoints==0.4
exceptiongroup==1.1.1
executing==0.8.3
fastapi==0.95.1
fastjsonschema==2.15.3
filelock==3.12.0
Flask==2.2.3
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
future==0.18.3
gcloud==0.18.3
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-core==2.3.2
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
h11==0.14.0
hnswlib==0.7.0
httplib2==0.22.0
httptools==0.5.0
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.4.1
importlib-resources==5.12.0
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
japanize-matplotlib==1.1.3
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.5.1
jupyter==1.0.0
jupyter-client==7.3.1
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
jws==0.1.3
kiwisolver==1.4.4
langchain==0.0.160
line-bot-sdk==2.4.2
lz4==4.3.2
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.3
mistune==0.8.4
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==3.1
nltk==3.8.1
notebook==6.4.11
numexpr==2.8.4
numpy==1.24.2
oauth2client==4.1.3
openai==0.27.6
openapi-schema-pydantic==1.2.4
outcome==1.2.0
packaging==21.3
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
posthog==3.0.1
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==4.22.3
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
Pygments==2.12.0
pyparsing==3.0.9
pypdf==3.8.1
Pyrebase4==4.6.0
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jwt==2.0.1
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
pyzmq==23.0.0
qtconsole==5.3.0
QtPy==2.1.0
regex==2023.5.5
requests==2.28.2
requests-toolbelt==0.10.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
selenium==4.9.0
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.12
stack-data==0.2.0
starlette==0.26.1
sympy==1.11.1
tenacity==8.2.2
terminado==0.15.0
threadpoolctl==3.1.0
tiktoken==0.3.3
tinycss2==1.1.1
tokenizers==0.13.3
torch==2.0.0
torchvision==0.15.1
tornado==6.1
tqdm==4.65.0
traitlets==5.2.1.post0
transformers==4.28.1
trio==0.22.0
trio-websocket==0.10.2
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.5
webdriver-manager==3.8.6
webencodings==0.5.1
websockets==11.0.2
Werkzeug==2.2.3
widgetsnbextension==3.6.0
wsproto==1.2.0
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0
